# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.AllocationSummary:
    properties:
      allocated:
        type: number
      category:
        type: string
    type: object
...
  handlers.CreateIncomeRequest:
    properties:
      account_id:
        type: string
      amount:
        type: number
      date:
        type: string
      description:
        type: string
      source:
        type: string
    required:
    - account_id
    - amount
    - date
    - source
    type: object
  handlers.CreateIncomeResponse:
    properties:
      allocations:
        items:
          $ref: '#/definitions/handlers.AllocationSummary'
        type: array
      income:
        $ref: '#/definitions/models.Income'
    type: object
...
  handlers.CreateTransferRequest:
    properties:
      amount:
        type: number
      date:
        type: string
      from_account_id:
        type: string
      note:
        type: string
      to_account_id:
        type: string
    required:
    - amount
    - date
    - from_account_id
    - to_account_id
    type: object
...
  models.Account:
    properties:
      balance:
        type: number
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      expenses:
        items:
          $ref: '#/definitions/models.Expense'
        type: array
      goal_amount:
        description: SAVINGS only
        type: number
      goal_label:
        type: string
      id:
        type: string
      income_type:
        description: 'Only for CARD: SALARY|PROJECT|FREELANCE|BUSINESS|OTHER'
        type: string
      incomes:
        description: Relations
        items:
          $ref: '#/definitions/models.Income'
        type: array
      is_active:
        type: boolean
      name:
        type: string
      transfers:
        items:
          $ref: '#/definitions/models.AccountTransfer'
        type: array
      type:
        description: CARD | CASH | SAVINGS
        type: string
      updated_at:
        type: string
    type: object
  models.AccountTransfer:
    properties:
      amount:
        type: number
      created_at:
        type: string
      from_account:
        allOf:
        - $ref: '#/definitions/models.Account'
        description: Relations
      from_account_id:
        type: string
      id:
        type: string
      note:
        type: string
      to_account:
        $ref: '#/definitions/models.Account'
      to_account_id:
        type: string
      transfer_date:
        type: string
      updated_at:
        type: string
    type: object
  models.BudgetAllocation:
    properties:
      allocated_amount:
        type: number
      category:
        $ref: '#/definitions/models.ExpenseCategory'
      category_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      income:
        allOf:
        - $ref: '#/definitions/models.Income'
        description: Relations
      income_id:
        type: string
      month:
        type: integer
      year:
        type: integer
    type: object
  models.CategoryBudget:
    properties:
      allocated_amount:
        type: number
      category:
        allOf:
        - $ref: '#/definitions/models.ExpenseCategory'
        description: Relations
      category_id:
        type: string
      created_at:
        type: string
      days_in_month:
        description: Snapshot of days in month
        type: integer
      effective_daily_amount:
        description: Snapshot for DAILY_CONTINUOUS
        type: number
      id:
        type: string
      month:
        description: 1-12
        type: integer
      remaining_amount:
        type: number
      spent_amount:
        type: number
      updated_at:
        type: string
      year:
        type: integer
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  models.Expense:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/models.Account'
        description: Relations
      account_id:
        description: Account used for payment
        type: string
      amount:
        type: number
      category:
        $ref: '#/definitions/models.ExpenseCategory'
      category_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  models.ExpenseCategory:
    properties:
      allocation_priority:
        type: integer
      allocations:
        items:
          $ref: '#/definitions/models.BudgetAllocation'
        type: array
      budgets:
        description: Relations
        items:
          $ref: '#/definitions/models.CategoryBudget'
        type: array
      created_at:
        type: string
      daily_amount:
        description: Only for DAILY_CONTINUOUS
        type: number
      expenses:
        items:
          $ref: '#/definitions/models.Expense'
        type: array
      id:
        type: string
      is_active:
        type: boolean
      metadata:
        type: object
      monthly_budget:
        type: number
      name:
        type: string
      type:
        description: SUBSCRIPTION, DAILY_CONTINUOUS, USAGE_BASED, ONE_TIME
        type: string
      updated_at:
        type: string
    type: object
  models.Income:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/models.Account'
        description: Relations
      account_id:
        description: Source account
        type: string
      allocations:
        items:
          $ref: '#/definitions/models.BudgetAllocation'
        type: array
      amount:
        type: number
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: string
      source:
        type: string
      updated_at:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@financeapp.com
    name: API Support
    url: https://github.com/yourusername/finance-tracking-app
  description: A comprehensive personal finance tracking application API with budget
    management, expense tracking, and analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Finance Tracking API
  version: "1.0"
paths:
...
  /incomes:
    get:
      description: Get all incomes with pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all incomes
      tags:
      - income
    post:
      consumes:
      - application/json
      description: Create a new income and automatically allocate budget to categories
      parameters:
      - description: Income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateIncomeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateIncomeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create new income
      tags:
      - income
  /incomes/{id}:
    delete:
      description: Delete income by ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete income
      tags:
      - income
    get:
      description: Get a specific income by ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get income by ID
      tags:
      - income
    patch:
      consumes:
      - application/json
      description: Update income by ID and re-calculate allocations
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateIncomeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateIncomeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update income
      tags:
      - income
...
  /transfers:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get all transfers
      tags:
      - transfers
    post:
      consumes:
      - application/json
      parameters:
      - description: Transfer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: Transfer funds between accounts
      tags:
      - transfers
  /transfers/{id}:
    delete:
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Cancel a transfer (within 24 hours)
      tags:
      - transfers
    get:
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get transfer by ID
      tags:
      - transfers
  /transfers/account:
    get:
      parameters:
      - description: Account ID
        in: query
        name: account_id
        required: true
        type: string
      - description: Month (1-12)
        in: query
        name: month
        type: integer
      - description: Year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get transfers for a specific account
      tags:
      - transfers
schemes:
- http
- https
swagger: "2.0"